% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/actvspred.R
\name{actvspred}
\alias{actvspred}
\title{Comparison of actual and predicted linear response}
\usage{
actvspred(y, yhat, n)
}
\arguments{
\item{y}{actual linear response}

\item{yhat}{predicted linear response}

\item{n}{quantiles to be created for comparison}
}
\value{
a data.frame with average actual and predicted response in each quantile
}
\description{
Takes in actual, predicted linear response and quantile value, and returns
 average actual and predicted response in each quantile
}
\examples{
# a data frame with y and yhat
df <- data.frame(y = c(1, 2, 3, 6, 8, 10, 15),
                 yhat = c(1.2, 2.5, 3.3, 6.9, 9.3, 6.5, 12.3))

# Get actual vs predicted table
ACTVSPRED <- actvspred(y = df[, 'y'], yhat = df[, 'yhat'], n = 5)
}
\author{
Akash Jain
}
\seealso{
\code{\link{mape}}, \code{\link{splitdata}}
}

