% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analytics_function.r
\name{Analytics_function}
\alias{Analytics_function}
\title{An Analytics Function}
\usage{
Analytics_function(love = TRUE)
}
\arguments{
\item{love}{Do you love data? Defaults to TRUE.}
}
\description{
This function allows you to express your love for data.
}
\examples{
Analytics_function()
}
\keyword{data}
