% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_and_clean_db.R
\name{archive_db}
\alias{archive_db}
\title{Archive Database}
\usage{
archive_db(user = "", password = "")
}
\arguments{
\item{user}{Username for the database}

\item{password}{Password to login for given username}
}
\value{
All tables in the database are copied and archived
}
\description{
Archive Database
}
\details{
This is a helper function to archive data before tables are deleted. Old archives are overwritten.
}
\examples{
\dontrun{
archive_db("janedoe", "areallystrongpassword")
}
}
