% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_list_named_queries}
\alias{athena_list_named_queries}
\title{Provides a list of available query IDs only for queries saved in the
specified workgroup}
\usage{
athena_list_named_queries(NextToken, MaxResults, WorkGroup)
}
\arguments{
\item{NextToken}{The token that specifies where to start pagination if a previous request
was truncated.}

\item{MaxResults}{The maximum number of queries to return in this request.}

\item{WorkGroup}{The name of the workgroup from which the named queries are being
returned.}
}
\description{
Provides a list of available query IDs only for queries saved in the
specified workgroup. Requires that you have access to the workgroup.
}
\details{
For code samples using the AWS SDK for Java, see \href{http://docs.aws.amazon.com/athena/latest/ug/code-samples.html}{Examples and Code Samples}
in the \emph{Amazon Athena User Guide}.
}
\section{Request syntax}{
\preformatted{athena$list_named_queries(
  NextToken = "string",
  MaxResults = 123,
  WorkGroup = "string"
)
}
}

\keyword{internal}
