% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_untag_resource}
\alias{athena_untag_resource}
\title{Removes one or more tags from the workgroup resource}
\usage{
athena_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] Removes one or more tags from the workgroup resource for the specified
ARN.}

\item{TagKeys}{[required] Removes the tags associated with one or more tag keys from the workgroup
resource.}
}
\description{
Removes one or more tags from the workgroup resource. Takes as an input
a list of TagKey Strings separated by commas, and removes their tags at
the same time.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
