% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_and_clean_db.R
\name{clean_and_archive_db}
\alias{clean_and_archive_db}
\title{Cleans and archives the remote Postgres Database}
\usage{
clean_and_archive_db(user = "", password = "")
}
\arguments{
\item{user}{Username for the database}

\item{password}{Password to login for given username}
}
\value{
Archives all existing tables (except admin) and deletes them to allow new data to be entered
}
\description{
Cleans and archives the remote Postgres Database
}
\details{
This is a helper function to clean and archive the database. It first archives all tables and then deletes them.
}
\examples{
\dontrun{
clean_and_archive_db("janedoe", "areallystrongpassword")
}
}
