% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_and_clean_db.R
\name{clean_db}
\alias{clean_db}
\title{Delete Tables from database for cleaning}
\usage{
clean_db(user = "", password = "")
}
\arguments{
\item{user}{Username for the database}

\item{password}{Password to login for given username}
}
\value{
Deletes all tables except admin generated and archives.
}
\description{
Delete Tables from database for cleaning
}
\details{
This is a helper function to clean the database. It deletes all tables except admin generated and archives.
}
\examples{
\dontrun{
clean_db("janedoe", "areallystrongpassword")
}
}
