% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{clean_survey_data}
\alias{clean_survey_data}
\title{Clean Survey Data}
\usage{
clean_survey_data(data, n_course_modules = 10)
}
\arguments{
\item{data}{A data frame containing the raw survey data.}

\item{n_course_modules}{Number of modules in the course. Default is 10.}
}
\value{
A cleaned data frame.
}
\description{
Cleans the raw survey data retrieved from the Google Sheet.
}
