% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dqcategorical.R
\name{dqcategorical}
\alias{dqcategorical}
\title{Data quality check of categorical variables}
\usage{
dqcategorical(data)
}
\arguments{
\item{data}{a data.frame or data.table}
}
\value{
a data.frame which contains the variable, category index, category,
         category frequency and percentage frequency of all factor variables
}
\description{
Takes in a data, and returns summary of categorical variables
}
\examples{
# Let's create a data.frame
df <- data.frame(phone = c('IP', 'SN', 'HO', 'IP', 'SN', 'IP', 'HO', 'SN', 'IP', 'SN'),
                 colour = c('black', 'blue', 'green', 'blue', 'black', 'silver', 'black',
                 'white', 'black', 'green'))

# Factorise categorical variables
df <- factorise(data = df, colNames = c('phone', 'colour'))

# Generate a data quality report of continuous variables
summaryCategorical <- dqcategorical(data = df)
}
\author{
Akash Jain
}
\seealso{
\code{\link{dqcontinuous}}, \code{\link{dqdate}}, \code{\link{contents}}
}

