% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dqcontinuous.R
\name{dqcontinuous}
\alias{dqcontinuous}
\title{Data quality check of continuous variables}
\usage{
dqcontinuous(data)
}
\arguments{
\item{data}{a data.frame or data.table}
}
\value{
a data.frame which contains the non-missing values, missing values,
         percentage of missing values, mimimum, mean, maximum, standard deviation,
         variance, percentiles and count of outliers of all integer and
         numeric variables
}
\description{
Takes in a data, and returns summary of continuous variables
}
\examples{
# Let's create a data.frame
df <- data.frame(x = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
                 y = c(22, NA, 66, 12, 78, 34, 590, 97, 56, 37))

# Generate a data quality report of continuous variables
summaryContinuous <- dqcontinuous(data = df)
}
\author{
Akash Jain
}
\seealso{
\code{\link{dqcategorical}}, \code{\link{dqdate}}, \code{\link{contents}}
}

