% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_describe_reserved_elasticsearch_instance_offerings}
\alias{elasticsearchservice_describe_reserved_elasticsearch_instance_offerings}
\title{Lists available reserved Elasticsearch instance offerings}
\usage{
elasticsearchservice_describe_reserved_elasticsearch_instance_offerings(
  ReservedElasticsearchInstanceOfferingId, MaxResults, NextToken)
}
\arguments{
\item{ReservedElasticsearchInstanceOfferingId}{The offering identifier filter value. Use this parameter to show only
the available offering that matches the specified reservation
identifier.}

\item{MaxResults}{Set this value to limit the number of results returned. If not
specified, defaults to 100.}

\item{NextToken}{NextToken should be sent in case if earlier API call produced result
containing NextToken. It is used for pagination.}
}
\description{
Lists available reserved Elasticsearch instance offerings.
}
\section{Request syntax}{
\preformatted{elasticsearchservice$describe_reserved_elasticsearch_instance_offerings(
  ReservedElasticsearchInstanceOfferingId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
