% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticsearchservice_operations.R
\name{elasticsearchservice_list_tags}
\alias{elasticsearchservice_list_tags}
\title{Returns all tags for the given Elasticsearch domain}
\usage{
elasticsearchservice_list_tags(ARN)
}
\arguments{
\item{ARN}{[required] Specify the \code{ARN} for the Elasticsearch domain to which the tags are
attached that you want to view.}
}
\description{
Returns all tags for the given Elasticsearch domain.
}
\section{Request syntax}{
\preformatted{svc$list_tags(
  ARN = "string"
)
}
}

\keyword{internal}
