% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_job_flows}
\alias{emr_describe_job_flows}
\title{This API is deprecated and will eventually be removed}
\usage{
emr_describe_job_flows(CreatedAfter, CreatedBefore, JobFlowIds,
  JobFlowStates)
}
\arguments{
\item{CreatedAfter}{Return only job flows created after this date and time.}

\item{CreatedBefore}{Return only job flows created before this date and time.}

\item{JobFlowIds}{Return only job flows whose job flow ID is contained in this list.}

\item{JobFlowStates}{Return only job flows whose state is contained in this list.}
}
\description{
This API is deprecated and will eventually be removed. We recommend you
use ListClusters, DescribeCluster, ListSteps, ListInstanceGroups and
ListBootstrapActions instead.
}
\details{
DescribeJobFlows returns a list of job flows that match all of the
supplied parameters. The parameters can include a list of job flow IDs,
job flow states, and restrictions on job flow creation date and time.

Regardless of supplied parameters, only job flows created within the
last two months are returned.

If no parameters are supplied, then job flows matching either of the
following criteria are returned:
\itemize{
\item Job flows created and completed in the last two weeks
\item Job flows created within the last two months that are in one of the
following states: \code{RUNNING}, \code{WAITING}, \code{SHUTTING_DOWN}, \code{STARTING}
}

Amazon EMR can return a maximum of 512 job flow descriptions.
}
\section{Request syntax}{
\preformatted{svc$describe_job_flows(
  CreatedAfter = as.POSIXct(
    "2015-01-01"
  ),
  CreatedBefore = as.POSIXct(
    "2015-01-01"
  ),
  JobFlowIds = list(
    "string"
  ),
  JobFlowStates = list(
    "STARTING"|"BOOTSTRAPPING"|"RUNNING"|"WAITING"|"SHUTTING_DOWN"|"TERMINATED"|"COMPLETED"|"FAILED"
  )
)
}
}

\keyword{internal}
