% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_specific_to_frontiers2017_dataset.r
\name{fix_last_posture_of_index_dfs}
\alias{fix_last_posture_of_index_dfs}
\title{81520 is the length of the last posture where y was fixed, via manual id by cohn}
\usage{
fix_last_posture_of_index_dfs(list_of_idx_df)
}
\arguments{
\item{idxs}{dataframe, that contain the coordinates of the points, and adept_x and adept_y columns, for (fixX, fixY)}
}
\value{
fixed_posture_indices now with rm posture 208, and added final point for last 1000th fixed-y posture.
}
\description{
81520 is the length of the last posture where y was fixed, via manual id by cohn
}
