% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gadataupload.R
\name{gaUploadData}
\alias{gaUploadData}
\title{Google Analytics data upload}
\usage{
gaUploadData(accountId = NULL, webPropertyId = NULL,
  customDataSourceId = NULL, filename = NULL, token)
}
\arguments{
\item{accountId}{integer or character. Account ID for the upload to retrieve.}

\item{webPropertyId}{character. Web property ID for the upload to retrieve, for example UA-123456-1.}

\item{customDataSourceId}{character. Custom data source ID to which the data being uploaded belongs.}

\item{filename}{character. Custom data filename along with path.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}
}
\value{
uploaded or failed

\enumerate{
 \item More information in references.
}
}
\description{
Google Analytics data upload
}
\references{
\href{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/uploads#resource}{Uploads Resources}

\href{https://developers.google.com/analytics/devguides/config/mgmt/v3/mgmtReference/management/uploads/list}{Uploads list Method}
}
\seealso{
\code{\link{oauth}}

Other Google Analytics Management API: \code{\link{listGaCustomDatasources}}
}

