% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plausible_data.R
\name{get_data}
\alias{get_data}
\title{Send Request to Plausible for Data}
\usage{
get_data(url, TOKEN)
}
\arguments{
\item{url}{A character string specifying the URL to send the request to.}

\item{TOKEN}{A character string specifying the authentication token.}
}
\value{
A list of results from the request.
}
\description{
Send Request to Plausible for Data
}
\examples{
\dontrun{
country_data <- get_data(country_url, token = "MY_TOKEN")
}
}
