% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plausible_data.R
\name{get_plausible_data}
\alias{get_plausible_data}
\title{Retrieve and Save Web Analytics Data from Plausible}
\usage{
get_plausible_data(
  site_url = "",
  pagewise = FALSE,
  token = Sys.getenv("PLAUSIBLE_TOKEN")
)
}
\arguments{
\item{site_url}{A character string specifying the website URL to retrieve data for. If left empty,
the default URL "ds4owd-001.github.io/website" is used.}

\item{pagewise}{A logical value indicating whether to retrieve pagewise breakdown data. Default is `FALSE`.}

\item{token}{A character string for the Plausible API token. Defaults to retrieving the token from
the environment variable "PLAUSIBLE_TOKEN". See the Readme for more information on how to save tokens
to the R environment.}
}
\value{
This function does not return a value. It saves the retrieved data as CSV files.
}
\description{
This function fetches analytics data from the Plausible API for a given website and saves it into CSV files.
The data includes metrics like pageviews, visitors, bounce rate, and more. The function allows customization
of the website URL and the option to retrieve pagewise data.
}
\details{
The function constructs API requests to Plausible to retrieve data for countries, sources, browsers,
and a time series of visits. It saves the data in CSV files within a folder named after the site.

The data is broken down by country, source, and browser. Optionally, if `pagewise = TRUE`, it also retrieves
pagewise data.

The time series data is for the period from 2023-09-27 (when the website was initially created) to the current date.

The resulting data is saved in the `data/` folder under a subdirectory named after the first part of the website URL.
}
\examples{
\dontrun {
get_plausible_data(site_url = "ds4owd-001.github.io/website", pagewise = TRUE, token = "MY_PLAUSIBLE_TOKEN")
}

}
