% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pscloud_data.R
\name{get_pscloud_data}
\alias{get_pscloud_data}
\title{Pull Data from Posit Cloud}
\usage{
get_pscloud_data(space_id = 426916)
}
\arguments{
\item{space_id}{ID of the space to retrieve data from. Defaults to the DS$OWD space ID.}
}
\value{
Saves a CSV file with daily usage data. The file is stored as /data/posit_cloud_data/pscloud.csv
}
\description{
Pull Data from Posit Cloud
}
\examples{
space_id <- 999999
get_pscloud_data(space_id)
}
