% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_registration_data.R
\name{get_registrations_data}
\alias{get_registrations_data}
\title{Title: Download and Load Data}
\usage{
get_registrations_data(
  repo = "ds4owd",
  download_token = Sys.getenv("GITHUB_DOWNLOAD_TOKEN"),
  github_token = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{download_token}{- See README on how to obtain this token}

\item{github_token}{- GitHub Personal Access Token (PAT) with 'repo' scope}
}
\value{
Exports RDA and CSV files for locations, presurvey, and roster data
}
\description{
Title: Download and Load Data
}
\examples{
download_and_load_data(download_token = DOWNLOAD_TOKEN, github_token = GITHUB_PAT)

}
