% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pscloud_data.R
\name{get_space_data}
\alias{get_space_data}
\title{Pull Data Posit Cloud}
\usage{
get_space_data(space_id = 426916)
}
\arguments{
\item{space_id}{ID of the space to retrieve data from. Defaults to the DS$OWD space ID.}
}
\value{
Saves a CSV file with daily usage data. The file is stored as /data-raw/space_member_daily_usage.csv
}
\description{
Pull Data Posit Cloud
}
\examples{
space_id <- 999999
get_space_data(space_id)
}
