% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Obtain Post-Course Survey Data from Google Sheets}
\usage{
get_survey_data(
  sheet_url = "",
  service_acc_path = "path/to/service/account.json",
  email = "jane.doe@gmail.com",
  n_course_modules = 10
)
}
\arguments{
\item{sheet_url}{URL to the Google Sheet. Defaults to the post-course survey for 2023.}

\item{service_acc_path}{Path to the JSON file of the service account with access to the Google Sheet. Not required if `email` is provided.}

\item{email}{Email address of the user with access to the Google Sheet. Defaults to a dummy email. Either this or `service_acc_path` is required.}

\item{n_course_modules}{Number of modules in the course. Default is 10.}
}
\value{
A CSV file with the cleaned survey data.
}
\description{
Authenticate using a Gmail account or a service account and retrieve data from the post-survey sheet for the DS4OWD course.
}
\examples{
\dontrun{
get_survey_data(sheet_url = "docs.google.com/spreadsheets/d/YOUR_SHEET_ID/edit",
                email = "jane.doe@gmail.com", n_course_modules = 10)
}
}
