% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Gini coefficient of a distribution}
\usage{
gini(y)
}
\arguments{
\item{y}{an integer or numeric distribution}
}
\value{
gini coefficient of the distribution
}
\description{
Takes in a distribution and returns gini coefficient
}
\examples{
# Distribution
dist <- c(1, 4, 7, 15, 10)

# Gini coefficient
GINI <- gini(y = dist)
}
\author{
Akash Jain
}
\seealso{
\code{\link{auc}}
}

