% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_dev_endpoint}
\alias{glue_delete_dev_endpoint}
\title{Deletes a specified DevEndpoint}
\usage{
glue_delete_dev_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{[required] The name of the DevEndpoint.}
}
\description{
Deletes a specified DevEndpoint.
}
\section{Request syntax}{
\preformatted{glue$delete_dev_endpoint(
  EndpointName = "string"
)
}
}

\keyword{internal}
