% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_crawlers}
\alias{glue_get_crawlers}
\title{Retrieves metadata for all crawlers defined in the customer account}
\usage{
glue_get_crawlers(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The number of crawlers to return on each call.}

\item{NextToken}{A continuation token, if this is a continuation request.}
}
\description{
Retrieves metadata for all crawlers defined in the customer account.
}
\section{Request syntax}{
\preformatted{svc$get_crawlers(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
