% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_trigger}
\alias{glue_get_trigger}
\title{Retrieves the definition of a trigger}
\usage{
glue_get_trigger(Name)
}
\arguments{
\item{Name}{[required] The name of the trigger to retrieve.}
}
\description{
Retrieves the definition of a trigger.
}
\section{Request syntax}{
\preformatted{glue$get_trigger(
  Name = "string"
)
}
}

\keyword{internal}
