% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_reset_job_bookmark}
\alias{glue_reset_job_bookmark}
\title{Resets a bookmark entry}
\usage{
glue_reset_job_bookmark(JobName)
}
\arguments{
\item{JobName}{[required] The name of the job in question.}
}
\description{
Resets a bookmark entry.
}
\section{Request syntax}{
\preformatted{svc$reset_job_bookmark(
  JobName = "string"
)
}
}

\keyword{internal}
