% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_update_job}
\alias{glue_update_job}
\title{Updates an existing job definition}
\usage{
glue_update_job(JobName, JobUpdate)
}
\arguments{
\item{JobName}{[required] Name of the job definition to update.}

\item{JobUpdate}{[required] Specifies the values with which to update the job definition.}
}
\description{
Updates an existing job definition.
}
\section{Request syntax}{
\preformatted{svc$update_job(
  JobName = "string",
  JobUpdate = list(
    Description = "string",
    LogUri = "string",
    Role = "string",
    ExecutionProperty = list(
      MaxConcurrentRuns = 123
    ),
    Command = list(
      Name = "string",
      ScriptLocation = "string"
    ),
    DefaultArguments = list(
      "string"
    ),
    Connections = list(
      Connections = list(
        "string"
      )
    ),
    MaxRetries = 123,
    AllocatedCapacity = 123,
    Timeout = 123,
    MaxCapacity = 123.0,
    NotificationProperty = list(
      NotifyDelayAfter = 123
    ),
    SecurityConfiguration = "string"
  )
)
}
}

\keyword{internal}
