% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_service.R
\name{kafka}
\alias{kafka}
\title{Managed Streaming for Kafka}
\usage{
kafka()
}
\description{
The operations for managing an Amazon MSK cluster.
}
\section{Operations}{

\tabular{ll}{
\link[=kafka_create_cluster]{create_cluster} \tab Creates a new MSK cluster \cr
\link[=kafka_delete_cluster]{delete_cluster} \tab Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in the request \cr
\link[=kafka_describe_cluster]{describe_cluster} \tab Returns a description of the MSK cluster whose Amazon Resource Name (ARN) is specified in the request\cr
\link[=kafka_get_bootstrap_brokers]{get_bootstrap_brokers} \tab A list of brokers that a client application can use to bootstrap \cr
\link[=kafka_list_clusters]{list_clusters} \tab Returns a list of clusters in an account \cr
\link[=kafka_list_nodes]{list_nodes} \tab Returns a list of the broker nodes in the cluster
}
}

\examples{
\donttest{svc <- kafka()
svc$create_cluster(
  Foo = 123
)}

}
