% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kafka_operations.R
\name{kafka_describe_cluster}
\alias{kafka_describe_cluster}
\title{Returns a description of the MSK cluster whose Amazon Resource Name
(ARN) is specified in the request}
\usage{
kafka_describe_cluster(ClusterArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}
}
\description{
Returns a description of the MSK cluster whose Amazon Resource Name
(ARN) is specified in the request.
}
\section{Request syntax}{
\preformatted{svc$describe_cluster(
  ClusterArn = "string"
)
}
}

\keyword{internal}
