% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis_operations.R
\name{kinesis_describe_stream_summary}
\alias{kinesis_describe_stream_summary}
\title{Provides a summarized description of the specified Kinesis data stream
without the shard list}
\usage{
kinesis_describe_stream_summary(StreamName)
}
\arguments{
\item{StreamName}{[required] The name of the stream to describe.}
}
\description{
Provides a summarized description of the specified Kinesis data stream
without the shard list.
}
\details{
The information returned includes the stream name, Amazon Resource Name
(ARN), status, record retention period, approximate creation time,
monitoring, encryption details, and open shard count.
}
\section{Request syntax}{
\preformatted{svc$describe_stream_summary(
  StreamName = "string"
)
}
}

\keyword{internal}
