% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listing_data_to_JSON}
\alias{listing_data_to_JSON}
\title{Transform a Vidual input data to JSON for visualization}
\usage{
listing_data_to_JSON(vimeo_youtube_URL = NULL, images = NULL,
  property_type = NULL, property_view = NULL, architectural_style = NULL,
  sale_or_rent = NULL, rent_period = NULL, listing_price = NULL,
  listing_date = NULL, beds = NULL, baths_full = NULL,
  baths_half = NULL, baths_total = NULL, neighborhood = NULL,
  address_1 = NULL, address_2 = NULL, city = NULL, county = NULL,
  state_province = NULL, country = NULL, zip_postal = NULL,
  mls_num = NULL, lot_size = NULL, square_feet = NULL,
  garage_parking_spaces = NULL, year_built = NULL, lon = NULL,
  lat = NULL, short_description = NULL, amenities = NULL)
}
\arguments{
\item{vimeo_youtube_URL}{}

\item{images}{}

\item{property_type}{}

\item{property_view}{}

\item{architectural_style}{}

\item{sale_or_rent}{}

\item{rent_period}{}

\item{listing_price}{}

\item{listing_date}{}

\item{beds}{}

\item{baths_full}{}

\item{baths_half}{}

\item{baths_total}{}

\item{neighborhood}{}

\item{address_1}{}

\item{address_2}{}

\item{city}{}

\item{county}{}

\item{state_province}{}

\item{country}{}

\item{zip_postal}{}

\item{mls_num}{}

\item{lot_size}{}

\item{square_feet}{}

\item{garage_parking_spaces}{}

\item{year_built}{}

\item{lon}{}

\item{lat}{}

\item{short_description}{}

\item{amenities}{}
}
\description{
This function takes a input parameters for a Vidual listing and outputs a JSON file for visualization.
}

