% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mturk_operations.R
\name{mturk_list_worker_blocks}
\alias{mturk_list_worker_blocks}
\title{The ListWorkersBlocks operation retrieves a list of Workers who are
blocked from working on your HITs}
\usage{
mturk_list_worker_blocks(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{Pagination token}

\item{MaxResults}{}
}
\description{
The \code{ListWorkersBlocks} operation retrieves a list of Workers who are
blocked from working on your HITs.
}
\section{Request syntax}{
\preformatted{svc$list_worker_blocks(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
