% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_summary.R
\name{network_summary}
\alias{network_summary}
\title{Calculates basic node summary statistics}
\usage{
network_summary(c)
}
\arguments{
\item{c}{an igraph object}
}
\value{
a data.frame of node summary statistics
}
\description{
Calculates basic node summary statistics
}
\examples{
adj_matrix <- round(matrix(data = rep(runif(100, min = 0, max = 1)), 10,10),1)
diag(adj_matrix) <- 0
sample_igraph <- igraph::graph_from_adjacency_matrix(adj_matrix)
network_summary_table <- network_summary(sample_igraph)
print(network_summary_table)

}
