% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_statistics.R
\name{node_statistics}
\alias{node_statistics}
\title{A wrapper function to more quickly calculate common node statistics on an igraph object}
\usage{
node_statistics(c)
}
\arguments{
\item{c}{an igraph object}
}
\value{
a dataframe containing one row for each node of an igraph with centrality measures
}
\description{
A wrapper function to more quickly calculate common node statistics on an igraph object
}
\examples{
adj_matrix <- round(matrix(data = rep(runif(100, min = 0, max = 1)), 10,10),1)
diag(adj_matrix) <- 0
sample_igraph <- igraph::graph_from_adjacency_matrix(adj_matrix)
node_statistics_df <- node_statistics(sample_igraph)
print(node_statistics)






}
