% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmdupobs.R
\name{rmdupobs}
\alias{rmdupobs}
\title{Remove duplicate observations from data}
\usage{
rmdupobs(data)
}
\arguments{
\item{data}{a data.frame or data.table}
}
\value{
a \code{data} of same class as input with only unique observations
}
\description{
Takes in a data, and returns it with duplicate observations removed
}
\examples{
# Let's create a data.frame
df <- data.frame(x = c(1, 2, 5, 1), y = c(3, 3, 1, 3))

# Remove duplicate observations from data
dfUnq <- rmdupobs(data = df)
}
\author{
Akash Jain
}
\seealso{
\code{\link{randomise}}, \code{\link{rmdupkey}}, \code{\link{factorise}}
}

