% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_dashboard}
\alias{run_dashboard}
\title{Run the Dashboard}
\usage{
run_dashboard(..., options = list(launch.browser = interactive()))
}
\arguments{
\item{...}{Currently not used.}

\item{options}{A \code{\link{list}} of named options that should be passed to the
\code{\link[shiny:shinyApp]{shiny::shinyAppDir()}} call.}
}
\value{
A \pkg{shiny} application object.
}
\description{
A wrapper for \code{\link[shiny:shinyApp]{shiny::shinyAppDir()}}.
}
\examples{
\dontrun{
run_dashboard()
}
}
\author{
N. Frerebeau
}
\concept{shiny apps}
