% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_functions.r
\name{split_time_series_by_piecewise_commands}
\alias{split_time_series_by_piecewise_commands}
\title{It finds all plateaus and creates a new section for each plateau}
\usage{
split_time_series_by_piecewise_commands(time_series)
}
\arguments{
\item{time_series}{A dataframe that has a column called command_M0.}
}
\description{
It finds all plateaus and creates a new section for each plateau
}
