% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db.R
\name{write_db}
\alias{write_db}
\title{Write to Database}
\usage{
write_db(user = "", password = "")
}
\arguments{
\item{user}{Username for the database}

\item{password}{Password for above username}
}
\value{
Writes all csv and rda files in the data folder to the database
}
\description{
Write to Database
}
\details{
Change the host, port and database name to connect to chosen database.
}
\examples{
\dontrun{
write_db("janedoe", "areallystrongpassword")
}

}
