% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db.R
\name{write_table}
\alias{write_table}
\title{Write To Database Table}
\usage{
write_table(file, data, con)
}
\arguments{
\item{file}{filepath of data being written}

\item{data}{data to be written to the database}

\item{con}{variable, connection to the Postgres database}
}
\value{
Writes the data to the database
}
\description{
Write To Database Table
}
\details{
This is a helper function to check if the table exists, if not it creates the table and writes to it.
If the table exists, it overwrites existing data.
}
\examples{
\dontrun{
write_table("data/survey_data/postsurvey.csv", data, con)
}
}
