% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcmreport.R
\name{runDcmReport}
\alias{runDcmReport}
\title{Get report configuration details from Doubleclick Campaign Manager}
\usage{
runDcmReport(profileId = NULL, reportId = NULL, token)
}
\arguments{
\item{profileId}{integer or character. Profile ID where the report resides.}

\item{reportId}{integer or character. Report ID within report builder section of DCM.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}
}
\value{
Report configuration details.

\enumerate{
 \item More information in references.
}
}
\description{
Get report configuration details from Doubleclick Campaign Manager
}
\references{
\href{https://developers.google.com/doubleclick-advertisers/reporting/v2.1/reports#resource}{DCM Report Resources}

\href{https://developers.google.com/doubleclick-advertisers/reporting/v2.1/reports/get}{DCM Report File Get Method}
}
\seealso{
\code{\link{oauth}}
}

